added multi sub support and action plan generation
