+++ tooling/aprl-automate/requirements.txt
@@ -5,3 +5,4 @@ azure-identity
 
 pandas
 argparse
+openpyxl
